renamed the components
