+++ packages/core/src/components/ProgressBars/ProgressCard.stories.tsx
@@ -15,7 +15,7 @@
  */
 
 import React from 'react';
-import { ProgressCard } from './ProgressCard';
+import { ProgressCard } from './GaugeCard';
 import { Grid } from '@material-ui/core';
 
 const linkInfo = { title: 'Go to XYZ Location', link: '#' };
